\name{GBT.generate.params.uniform.step1}
\alias{GBT.generate.params.uniform.step1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ GBT.generate.params.uniform.step1
}
\description{
GBT.generate.params.uniform.step1
}
\usage{
GBT.generate.params.uniform.step1(par.range)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par.range}{
GBT.generate.params.uniform.step1
}
}
\details{
GBT.generate.params.uniform.step1
}
\value{
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Dario Solari
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
